refactor: move streak, XP, and daily challenge logic  from user_progress_dashboard to streak_xp.dart [Feature/Task] Refactor misplaced code into a new Dart file uprm-inso4101-2025-2026-s1/semester-project-pronunciation-coach issue#113
